requirement.txt
asgiref==3.4.1
Django==3.2.15
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.3
typing-extensions==4.1.1
